Moved ReviewModel to a separate file
