+++ jobs/appstore_rate.rb
@@ -1,25 +1,7 @@
 require 'net/http'
 require 'json'
 
-module AppStore
-
-	class ReviewModel
-		attr_reader :author_name,
-								:rating,
-								:title,
-								:text
-
-		def initialize(hash)
-			@author_name = hash['author']['name']['label']
-			@rating = hash['im:rating']['label']
-			@title = hash['title']['label']
-			@text = hash['content']['label']
-		end
-	end
-
-
-
-end
+require_relative '../lib/appstore/review_model'
 
 SCHEDULER.every '1m', :first_in => 0 do |job|
